libcamera: software_isp: Add debayerRaw10PLine() function
Move the inner-loop of debayerRaw10P() to debayerRaw10PLine() which
debayers 1 line of bayer-data.

Putting the inner-loop in a separate function allows to compiler to
optimize it better, but this is mainly a preparation patch for adding
specialized functions for decoding even and odd lines to optimize
debayering.

Tested on a X1 yoga gen 8 (brightness 162/400) with IPU6 + ov2740 sensor.
Power consumption measured with wifi/bt off and only running qcam, using:
watch -n 10 cat /sys/class/power_supply/BAT0/power_now
and then waiting for the power to stabilize.

Before this change: ~19W
After this change: ~19W

Signed-off-by: Hans de Goede <hdegoede@redhat.com>