libcamera: software_isp: Add gamma correction
Add gamma-correction, this uses a fixed gamma-correction value of 0.5
for now.

The gamma-correction is stored in the green color lookup table,
the green lookup table values are multiplied by the red / blue gains
when calculating the red / blue lookup tables.

Since we already have the lookup tables adding gamma correction
like this is a free operation.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>