libcamera: debayer_cpu: Stop requiring 2 lines of padding at the top/bottom
Since the standard bayer pattern debayer functions no take an array
of 3 src pointers (previous-, current- and next-line) we can easily
avoid the need to have 2 lines (2 to keep the pattern unchanged)
padding at the top / bottom.

Instead when debayering the first line, pass the next-line pointer also
as previous-line pointer and pass the prev-line pointer as next-line
pointer when debayering the last line.

This allows the output to be 4 lines higher, which is esp. useful
with the current hi556 kernel driver which has a max-height for
the bauer data of 722 lines. So before this change it was not
possible to generate a standard 1280x720 output.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>