pipeline: raspberrypi: vc4: Use the CameraSensor embedded data API
Now that the CameraSensor class implements support for embedded data,
use the corresponding API instead of retrieving the embedded data format
from the subdev manually.

This changes requires sensor kernel drivers that implement the upstream
embedded data API based on V4L2 streams. As the API is under development
and not merged in the upstream kernel yet, this breaks compatibility
with the downstream Raspberry Pi kernel.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>