Add atomisp pipeline handler
The atomisp driver's Android heritage makes it weird in that
even though it uses MC + subdev-s it is designed to primarily
be controlled through its /dev/video# node.

This initial pipeline-handler makes use of this by copying
the uvcvideo pipeline-handler which also directly operates
on the /dev/video# node as a starting point.

This copy of the uvcvideo pipeline-handler has the mapping
of v4l2-controls and USB (location) handling removed +
some small other changes to make things work with the atomisp.

Future work will enumerate the actual sensors and hookup
auto-exposure/-gain control using the controls on the
sensor v4l-subdev and auto-whitebalance control using
the color gains on the ISP v4l-subdev.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>