libcamera: Add camera_ member to CameraConfiguration
Almost all pipeline-handlers CameraConfiguration implementations need
a pointer back to the camera-data (their Camera::Private derived class)
for their validate() implementation.

But the lifetime of the camera-data is tied to the lifetime of the Camera
object and a CameraConfiguration object could out-live the Camera object
resulting in a dangling pointer.

To avoid this the CameraConfiguration implementation should take
a reference on the Camera shared_ptr. But many pipeline handlers get
this wrong at the moment and only store a pointer to the camera-data.

Add a std::shared_ptr<Camera> camera_ member to the CameraConfiguration
base-class so that all CameraConfiguration implementations always have
a reference to the Camera they belong to, to fix these lifetime issues
for all pipeline handlers.

FIXME: only simple, uvcvideo and ipu3 pipelines have been adjusted
for now!

Signed-off-by: Hans de Goede <hdegoede@redhat.com>