libcamera: ipa_manager: createIPA: Allow passing an IPA name to match
Parts of the software ISP and the ipa_soft_simple.so IPA may be useful for
more then 1 pipeline handler.

Currently createIPA() / IPAManager::module() assume that there is a 1:1
relationship between pipeline handlers and IPAs and IPA matching is done
based on pipe->name().

Add an optional ipaName argument to createIPA() which when set overrides
pipe->name(), allowing pipeline handlers to request a different IPA.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>