pipeline: simple: Fix delayedCtrls handling
ATM the simple pipeline-handler tries to apply sensor controls coming from
the IPA twice:

a) It connects DelayedControls::applyControls to the video_->frameStart
   signal.

b) It directly applies controls from SimpleCameraData::setSensorControls()

a) Cannot work for 2 reasons:

1. video_ is the /dev/video# capture node and V4L2_EVENT_FRAME_SYNC events
are emitted by the /dev/v4l-subdev# node for the CSI receiver

2. The frameStart signal will only be emitted after a successful call to
setFrameStartEnabled() on the CSI V4L2Subdevice.

Drop the non working connecting of video_->frameStart to
DelayedControls::applyControls and add comment to
SimpleCameraData::setSensorControls() explaining why the controls are
applied there and add a TODO comment about ideally using the
frameStart signal for this when available.

Also add a missing delayedCtrls_->reset() call to start() so that the old
control history is cleared when starting to stream a second time.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>