libcamera: software_isp: Add valid flag to struct SwIspStats
Generating statistics for every single frame is not really necessary.

However a roundtrip through ipa_->processStats() still need to be done
every frame, even if there are no stats to make the IPA generate metadata
for every frame.

Add a valid flag to the statistics struct to let the IPA know when there
are no statistics for the frame being processed and modify the IPA to
only generate metadata for frames without valid statistics.

This is a preparation patch for skipping statistics generation for some
frames.

Signed-off-by: Hans de Goede <hansg@kernel.org>
---
Changes in v3:
- Drop if (!stats_->valid) early exit from IPASoftSimple::processStats()
- Save last AGC calculated new gain and exposure values and re-use these
  for frames without statistics, this avoids delayedCtrl history underruns
- Improve SwIspStats.valid documentation

Changes in v2:
- This is a new patch in v2 of this series