qcam: Sync buffers before import
Use the DmaSyncer handler to ensure the data from the buffer is
coherent. This is required by the spec - from
https://docs.kernel.org/driver-api/dma-buf.html#c.dma_buf_sync:

> When a DMA buffer is accessed from the CPU via mmap, it is not always
> possible to guarantee coherency between the CPU-visible map and
> underlying memory. To manage coherency, DMA_BUF_IOCTL_SYNC must be used
> to bracket any CPU access to give the kernel the chance to shuffle memory
> around if needed.

This was reported to fix glitches with the upcoming GPU-ISP, in which
case the accessed data in written by the GPU.

Note that in the GPU-ISP case we are effectively seeing a round-trip of
the buffer contents - from GPU synced to CPU, copied to GPU again. We
could avoid that in the future by implementing direct dmabuf import in
qcam.

Signed-off-by: Robert Mader <robert.mader@collabora.com>
Signed-off-by: Hans de Goede <johannes.goede@oss.qualcomm.com>