libcamera: shaders: Add support for black level compenstation
Add the ability to apply a int blacklevel gain against the demosaiced data
prior to application of colour correction data.

We calculate the index into the RGB black-level table so that we can apply
the black level when we sample the input prior to the demosiac operation.

Don't do this when using the lookup tables as the black level is
already calculated in those.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>