libcamera: software_isp: Implement a static init() routine
This is an overloaded init() routine that allows the CCM class to self
enumerate a default identity CCM at colour temperature 6500k.

This is required for the case where we are running GPUISP but don't have a
CCM for the sensor. In this case we want to generate a default CCM and use
it instead of using the CPUISP's lookup tables.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>